request config from the index file
